Use big decimals

+++ core/balance/src/main/kotlin/com/example/currencyconverter/core/balance/local/database/model/CurrencyBalanceEntity.kt
@@ -3,11 +3,12 @@ package com.example.currencyconverter.core.balance.local.database.model
 import androidx.room.ColumnInfo
 import androidx.room.Entity
 import com.example.currencyconverter.core.exchange.rates.model.Currency
+import java.math.BigDecimal
 
 @Entity(tableName = CurrencyBalanceEntity.TABLE_NAME, primaryKeys = ["currency"])
 internal data class CurrencyBalanceEntity(
     @ColumnInfo(name = COLUMN_CURRENCY) val currency: Currency,
-    @ColumnInfo(name = COLUMN_AMOUNT) val amount: Double,
+    @ColumnInfo(name = COLUMN_AMOUNT) val amount: BigDecimal,
 ) {
     companion object {
         const val TABLE_NAME = "currency_balance"
