+++ core/exchange/src/main/kotlin/com/example/currencyconverter/core/exchange/fees/FeeResolver.kt
@@ -1,11 +1,12 @@
 package com.example.currencyconverter.core.exchange.fees
 
 import com.example.currencyconverter.core.exchange.converter.model.ExchangeTransaction
+import java.math.BigDecimal
 
 internal interface FeeResolver {
 
     /**
      * Calculates fee amount for provided [ExchangeTransaction].
      */
-    suspend fun resolve(exchangeTransaction: ExchangeTransaction): Double
+    suspend fun resolve(exchangeTransaction: ExchangeTransaction): BigDecimal
 }
