+++ core/exchange/src/main/kotlin/com/example/currencyconverter/core/exchange/rates/model/ExchangeRate.kt
@@ -1,6 +1,8 @@
 package com.example.currencyconverter.core.exchange.rates.model
 
+import java.math.BigDecimal
+
 data class ExchangeRate(
     val currency: Currency,
-    val value: Double,
+    val value: BigDecimal,
 )
